refactor(team): extract DM slug helpers from broker.go
