+++ internal/team/broker.go
@@ -5,6 +5,7 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -13,7 +14,7 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"regexp"
+	"path/filepath"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -199,52 +200,8 @@ type teamChannel struct {
 	UpdatedAt   string          `json:"updated_at,omitempty"`
 }
 
-func (ch *teamChannel) isDM() bool {
-	return ch.Type == "dm" || IsDMSlug(ch.Slug)
-}
-
-// IsDMSlug checks whether a channel slug represents a direct message.
-func IsDMSlug(slug string) bool {
-	slug = normalizeChannelSlug(slug)
-	return strings.HasPrefix(slug, "dm-") || canonicalDMTargetAgent(slug) != ""
-}
-
-// DMSlugFor returns the DM channel slug for a given agent.
-func DMSlugFor(agentSlug string) string {
-	agentSlug = normalizeActorSlug(agentSlug)
-	if agentSlug == "" {
-		return ""
-	}
-	return channel.DirectSlug("human", agentSlug)
-}
-
-// DMTargetAgent extracts the agent slug from a DM channel slug.
-// Returns "" if the slug is not a DM.
-func DMTargetAgent(slug string) string {
-	slug = normalizeChannelSlug(slug)
-	if strings.HasPrefix(slug, "dm-human-") {
-		return strings.TrimPrefix(slug, "dm-human-")
-	}
-	if strings.HasPrefix(slug, "dm-") {
-		return strings.TrimPrefix(slug, "dm-")
-	}
-	return canonicalDMTargetAgent(slug)
-}
-
-func canonicalDMTargetAgent(slug string) string {
-	parts := strings.Split(normalizeChannelSlug(slug), "__")
-	if len(parts) != 2 {
-		return ""
-	}
-	switch {
-	case parts[0] == "human" || parts[0] == "you":
-		return parts[1]
-	case parts[1] == "human" || parts[1] == "you":
-		return parts[0]
-	default:
-		return ""
-	}
-}
+// DM slug helpers (isDM, IsDMSlug, DMSlugFor, DMTargetAgent,
+// canonicalDMTargetAgent) moved to broker_dm.go.
 
 type officeMember struct {
 	Slug           string                   `json:"slug"`
@@ -1947,309 +1904,13 @@ func (b *Broker) MemberProviderKind(slug string) string {
 }
 
 // memberFromSpec builds an officeMember from a manifest MemberSpec, threading
-// Provider through. Used by defaultOfficeMembers and by HTTP create paths so
-// field-copy logic lives in one place.
-func memberFromSpec(spec company.MemberSpec, createdBy, createdAt string, builtIn bool) officeMember {
-	return officeMember{
-		Slug:           spec.Slug,
-		Name:           spec.Name,
-		Role:           spec.Role,
-		Expertise:      append([]string(nil), spec.Expertise...),
-		Personality:    spec.Personality,
-		PermissionMode: spec.PermissionMode,
-		AllowedTools:   append([]string(nil), spec.AllowedTools...),
-		CreatedBy:      createdBy,
-		CreatedAt:      createdAt,
-		BuiltIn:        builtIn,
-		Provider:       spec.Provider,
-	}
-}
-
-// mentionPattern matches @slug tokens in free-form message text. Slugs are
-// alphanumeric plus hyphen, 2–30 chars (to dodge false positives from
-// conversational @ use). A preceding word character (e.g. email "a@b.com")
-// is excluded via the lookahead-free alternative: we anchor to start of
-// string or a non-alphanumeric rune.
-var mentionPattern = regexp.MustCompile(`(?:^|[^a-zA-Z0-9_])@([a-z0-9][a-z0-9-]{1,29})\b`)
-
-// extractMentionedSlugs pulls @-mention slugs out of body content. Duplicates
-// are removed. The caller is responsible for validating whether each slug is
-// a real office member.
-func extractMentionedSlugs(content string) []string {
-	matches := mentionPattern.FindAllStringSubmatch(strings.ToLower(content), -1)
-	if len(matches) == 0 {
-		return nil
-	}
-	seen := make(map[string]struct{}, len(matches))
-	out := make([]string, 0, len(matches))
-	for _, m := range matches {
-		if len(m) < 2 {
-			continue
-		}
-		slug := m[1]
-		if _, ok := seen[slug]; ok {
-			continue
-		}
-		seen[slug] = struct{}{}
-		out = append(out, slug)
-	}
-	return out
-}
-
-func uniqueSlugs(items []string) []string {
-	seen := make(map[string]struct{}, len(items))
-	out := make([]string, 0, len(items))
-	for _, item := range items {
-		item = normalizeChannelSlug(item)
-		if item == "" {
-			continue
-		}
-		if _, ok := seen[item]; ok {
-			continue
-		}
-		seen[item] = struct{}{}
-		out = append(out, item)
-	}
-	return out
-}
-
-func normalizeStringList(items []string) []string {
-	seen := make(map[string]struct{}, len(items))
-	out := make([]string, 0, len(items))
-	for _, item := range items {
-		item = strings.TrimSpace(item)
-		if item == "" {
-			continue
-		}
-		if _, ok := seen[item]; ok {
-			continue
-		}
-		seen[item] = struct{}{}
-		out = append(out, item)
-	}
-	return out
-}
-
-func containsString(items []string, want string) bool {
-	for _, item := range items {
-		if item == want {
-			return true
-		}
-	}
-	return false
-}
-
-func humanizeSlug(slug string) string {
-	parts := strings.Split(strings.ReplaceAll(strings.TrimSpace(slug), "-", " "), " ")
-	for i, part := range parts {
-		if part == "" {
-			continue
-		}
-		parts[i] = strings.ToUpper(part[:1]) + part[1:]
-	}
-	return strings.Join(parts, " ")
-}
-
-func defaultTeamChannelDescription(slug, name string) string {
-	manifest, err := company.LoadManifest()
-	if err == nil {
-		for _, ch := range manifest.Channels {
-			if normalizeChannelSlug(ch.Slug) == normalizeChannelSlug(slug) && strings.TrimSpace(ch.Description) != "" {
-				return strings.TrimSpace(ch.Description)
-			}
-		}
-	}
-	if normalizeChannelSlug(slug) == "general" {
-		return "The default company-wide room for top-level coordination, announcements, and cross-functional discussion."
-	}
-	label := strings.TrimSpace(name)
-	if label == "" {
-		label = humanizeSlug(slug)
-	}
-	return label + " focused work. Use this channel for discussion, decisions, and execution specific to that stream."
-}
-
-// reservedChannelSlugs are slug values that canAccessChannelLocked treats as
-// universally trusted senders. Any user-created channel sharing one of these
-// slugs would inherit that trust — every actor in the trust list could read
-// every message in that channel without an explicit Members entry. The
-// channel-create handler guards against this by rejecting create requests
-// whose slug matches this set; keep the two lists in sync.
-var reservedChannelSlugs = map[string]bool{
-	"system": true,
-	"nex":    true,
-	"you":    true,
-	"human":  true,
-}
-
-func (b *Broker) canAccessChannelLocked(slug, channel string) bool {
-	slug = normalizeActorSlug(slug)
-	channel = normalizeChannelSlug(channel)
-	if b.sessionMode == SessionModeOneOnOne {
-		if slug == "" || slug == "you" || slug == "human" {
-			return true
-		}
-		return slug == b.oneOnOneAgent
-	}
-	// NOTE: any new entry added here MUST also be added to
-	// reservedChannelSlugs above so the channel-create handler keeps the
-	// invariant "no user channel can shadow a trusted sender slug".
-	if slug == "" || slug == "you" || slug == "human" || slug == "nex" || slug == "system" {
-		return true
-	}
-	if slug == "ceo" {
-		return true
-	}
-	return b.channelHasMemberLocked(channel, slug)
-}
-
-func truncateSummary(s string, max int) string {
-	s = strings.TrimSpace(s)
-	if len([]rune(s)) <= max {
-		return s
-	}
-	runes := []rune(s)
-	if max <= 1 {
-		return "…"
-	}
-	return string(runes[:max-1]) + "…"
-}
-
-func applyOfficeMemberDefaults(member *officeMember) {
-	if member == nil {
-		return
-	}
-	if member.Name == "" {
-		member.Name = humanizeSlug(member.Slug)
-	}
-	if member.Role == "" {
-		member.Role = member.Name
-	}
-	if len(member.Expertise) == 0 {
-		member.Expertise = inferOfficeExpertise(member.Slug, member.Role)
-	}
-	if member.Personality == "" {
-		member.Personality = inferOfficePersonality(member.Slug, member.Role)
-	}
-	if member.PermissionMode == "" {
-		member.PermissionMode = "plan"
-	}
-}
-
-func inferOfficeExpertise(slug, role string) []string {
-	text := strings.ToLower(strings.TrimSpace(slug + " " + role))
-	switch {
-	case strings.Contains(text, "front"), strings.Contains(text, "ui"), strings.Contains(text, "design eng"):
-		return []string{"frontend", "UI", "interaction design", "components", "accessibility"}
-	case strings.Contains(text, "back"), strings.Contains(text, "api"), strings.Contains(text, "infra"):
-		return []string{"backend", "APIs", "systems", "infrastructure", "databases"}
-	case strings.Contains(text, "ai"), strings.Contains(text, "ml"), strings.Contains(text, "llm"):
-		return []string{"AI", "LLMs", "agents", "retrieval", "evaluations"}
-	case strings.Contains(text, "market"), strings.Contains(text, "brand"), strings.Contains(text, "growth"):
-		return []string{"marketing", "growth", "positioning", "campaigns", "brand"}
-	case strings.Contains(text, "revenue"), strings.Contains(text, "sales"), strings.Contains(text, "cro"):
-		return []string{"sales", "revenue", "pipeline", "partnerships", "closing"}
-	case strings.Contains(text, "product"), strings.Contains(text, "pm"):
-		return []string{"product", "roadmap", "requirements", "prioritization", "scope"}
-	case strings.Contains(text, "design"):
-		return []string{"design", "UX", "visual systems", "prototyping", "brand"}
-	default:
-		return []string{strings.ToLower(strings.TrimSpace(role))}
-	}
-}
-
-func inferOfficePersonality(slug, role string) string {
-	text := strings.ToLower(strings.TrimSpace(slug + " " + role))
-	switch {
-	case strings.Contains(text, "front"):
-		return "Frontend specialist focused on polished user-facing work and sharp interaction details."
-	case strings.Contains(text, "back"):
-		return "Systems-minded engineer who keeps complexity under control and worries about reliability."
-	case strings.Contains(text, "ai"), strings.Contains(text, "ml"), strings.Contains(text, "llm"):
-		return "AI engineer who likes ambitious ideas but immediately asks how they will actually work."
-	case strings.Contains(text, "market"), strings.Contains(text, "brand"), strings.Contains(text, "growth"):
-		return "Growth and positioning operator who translates product work into market momentum."
-	case strings.Contains(text, "revenue"), strings.Contains(text, "sales"):
-		return "Commercial operator who thinks in demand, objections, and revenue consequences."
-	case strings.Contains(text, "product"), strings.Contains(text, "pm"):
-		return "Product thinker who turns ambiguity into scope, sequencing, and crisp tradeoffs."
-	case strings.Contains(text, "design"):
-		return "Taste-driven designer who cares about clarity, craft, and how the product actually feels."
-	default:
-		return "A sharp teammate with a clear specialty, strong point of view, and enough personality to feel human."
-	}
-}
-
-func (b *Broker) channelHasMemberLocked(channel, slug string) bool {
-	ch := b.findChannelLocked(channel)
-	if ch == nil {
-		// Fall back to channelStore for new-format channels (e.g. "eng__human")
-		if b.channelStore != nil {
-			return b.channelStore.IsMemberBySlug(channel, slug)
-		}
-		return false
-	}
-	for _, member := range ch.Members {
-		if member == slug {
-			return true
-		}
-	}
-	return false
-}
-
-func (b *Broker) channelMemberEnabledLocked(channel, slug string) bool {
-	if !b.channelHasMemberLocked(channel, slug) {
-		return false
-	}
-	ch := b.findChannelLocked(channel)
-	if ch == nil {
-		return false
-	}
-	for _, disabled := range ch.Disabled {
-		if disabled == slug {
-			return false
-		}
-	}
-	return true
-}
-
-func (b *Broker) enabledChannelMembersLocked(channel string, candidates []string) []string {
-	var out []string
-	for _, candidate := range candidates {
-		if b.channelMemberEnabledLocked(channel, candidate) {
-			out = append(out, candidate)
-		}
-	}
-	return out
-}
-
-func (b *Broker) ensureTaskOwnerChannelMembershipLocked(channel, owner string) {
-	channel = normalizeChannelSlug(channel)
-	owner = normalizeChannelSlug(owner)
-	if channel == "" || owner == "" {
-		return
-	}
-	if b.findMemberLocked(owner) == nil {
-		return
-	}
-	ch := b.findChannelLocked(channel)
-	if ch == nil {
-		return
-	}
-	if !containsString(ch.Members, owner) {
-		ch.Members = uniqueSlugs(append(ch.Members, owner))
-		ch.UpdatedAt = time.Now().UTC().Format(time.RFC3339)
-	}
-	if len(ch.Disabled) > 0 {
-		filtered := ch.Disabled[:0]
-		for _, disabled := range ch.Disabled {
-			if disabled != owner {
-				filtered = append(filtered, disabled)
-			}
-		}
-		ch.Disabled = filtered
-	}
-}
+// Pure helpers + channel-membership predicates (memberFromSpec,
+// extractMentionedSlugs, uniqueSlugs, normalizeStringList, containsString,
+// humanizeSlug, defaultTeamChannelDescription, canAccessChannelLocked,
+// truncateSummary, applyOfficeMemberDefaults, inferOfficeExpertise,
+// inferOfficePersonality, channelHasMemberLocked,
+// channelMemberEnabledLocked, enabledChannelMembersLocked,
+// ensureTaskOwnerChannelMembershipLocked) moved to broker_helpers.go.
 
 func usageStateIsZero(state teamUsageState) bool {
 	if state.Total.TotalTokens > 0 || state.Total.CostUsd > 0 || state.Total.Requests > 0 {
